Rename paths.hpp -> filesystem.hpp
