+++ tests/codeowners.t.cpp
@@ -1,5 +1,5 @@
 #include "codeowners/codeowners.hpp"
-#include "codeowners/paths.hpp"
+#include "codeowners/filesystem.hpp"
 
 #include <gtest/gtest.h>
